refactor: 디스코드 패키지 이동 및 인터페이스 시그니처 반영
