+++ backend/src/main/java/com/bether/bether/connection/discord/infrastructure/DiscordSlashCommandListener.java
@@ -1,4 +1,4 @@
-package com.bether.bether.connection.discord.application.service;
+package com.bether.bether.connection.discord.infrastructure;
 
 import java.util.Objects;
 import lombok.extern.slf4j.Slf4j;
